ændret game til at implementere ruleset
